Extract spec helper methods for reuse
