+++ spec/features/visitors/anyone_can_complete_a_journey_spec.rb
@@ -21,24 +21,6 @@ feature "Anyone can start a journey" do
     expect(page).to have_content("Not started")
   end
 
-  def click_first_link_in_task_list
-    within("ol.app-task-list") do
-      first("a").click
-    end
-  end
-
-  def start_journey_from_category_and_go_to_question(category:)
-    stub_contentful_category(
-      fixture_filename: category
-    )
-
-    visit root_path
-
-    click_on(I18n.t("generic.button.start"))
-
-    click_first_link_in_task_list
-  end
-
   scenario "an answer must be provided" do
     start_journey_from_category_and_go_to_question(category: "radio-question.json")
